Add basic home page and login page.
Include .css file with basic styling for the page, including the header for
each page.

Setup basic login functionality to be expanded upon once the database is
established.

Write functions to print the header, check a user given password, show the
login box, and return the user to their prior page or to the home page.