Implement weight bracket interface skeleton.
Make tweaks to the style file to improve presentation quality.

Add weight.php and implement initial skeleton for the weight bracket interface
to include options on whether to change the price for an existing bracket
or add a new one. TODO: add an option to remove a weight bracket.

Add a function to prevent unauthorized access to given web pages.